requirements.txt
playwright
behave
allure-behave
